feat: Create MetricsPatch modules to hold metrics
Like the Metrics ConfigurationPatch, include the MetricsPatch modules if
Metrics is enabled/present
